Run the Shiny Application

This function should be run only after you have created the configuration file and placed all required files in the app folder. See vignette("quickstart", package = "shinyExprPortal") for help with setup or vignette("fullguide", package = "shinyExprPortal") for a complete configuration guide.

Usage

run_app(
  config_file,
  data_folder = "",
  custom_modules = NULL,
  nthreads = 1L,
  ...
)

Arguments

config_file

The name of the yaml configuration file

data_folder

Optional directory prefix for data files. Use this argument if you want to version your files across different folders

custom_modules

Optional list of available custom modules. See the 'Details' section.

nthreads

Optional number of threads/cores to speed up loading files and computing correlations on UNIX-based systems. Default is 1

...

Further optional arguments.

Value

Runs the app

Details

custom_modules should contain a list of names for user-defined modules that are loaded in the environment before calling run_app. Each module should be accompanied by the corresponding mod_moduleName_ui, mod_moduleName_server moduleName_config functions. These functions could be placed in a custom_modules.R file, for example, and loaded using source. The package will then parse the configuration file, and if it contains one of the custom module names, it will call the module configuration parsing function and add it to the interface. See vignette("customization") for a complete example.

Please note that if running on Windows, nthreads will be always set to 1 due to limitations on the current implementation.

See also

Examples

if (interactive()) {
run_app("config.yaml", nthreads = 4)
}